refactor: useDeleteUser
